requirements.txt
tensorflow_gpu
numpy
opencv-python
shapely
